% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCPCD.R
\name{AddCodesCPCD}
\alias{AddCodesCPCD}
\title{Lists CP/CD issuers and country/sector codes}
\usage{
AddCodesCPCD(data, update_info = TRUE)
}
\value{
List of issuers and country/sector.
}
\description{
Adds information about country and sector from
the newest data entries to the historical data.
}
