% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMaturityBuckets.R
\name{AddMaturityBuckets}
\alias{AddMaturityBuckets}
\title{Adds maturity categories}
\usage{
AddMaturityBuckets(data)
}
\arguments{
\item{data}{(CP/CD) data set.}
}
\description{
Adds maturity categories
}
\details{
\code{AddMaturityBuckets} legger til en kolonne med løpetidsbuckets for CP/CD-data fra DTCC
}
\examples{
\dontrun{
dtcc_data = dtcc_data \%>\% AddMaturityBuckets()
}
}
