% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCPCD.R
\name{EstimateIssueDateCPCD}
\alias{EstimateIssueDateCPCD}
\title{Plots CP/CD on top of LIBOR}
\usage{
EstimateIssueDateCPCD(data)
}
\arguments{
\item{data}{Data set to plot.}
}
\value{
Plot of 3M Libor with CP/CD transactions of
comparable maturity (80-100 days) on top.
}
\description{
Plots the interest rate of issued CP/CDs on
top of 3M Libor.
}
\examples{
PlotCPCDAgainstLIBOR(FastLoadCPCD("2020-01-01", "2020-06-01"))
PlotCPCDAgainstLIBOR(dtcc_df)
}
