% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCPCD.R
\name{FormatDataCPCD}
\alias{FormatDataCPCD}
\title{Formats CP/CD data}
\usage{
FormatDataCPCD(data, update_info = FALSE)
}
\arguments{
\item{data}{(CP/CD) data set.}
}
\value{
Formatted CP/CD data.
}
\description{
Formats CP/CD data. Adds additional columns
that indicate (1) whether it is an issuance and
(2) whether the issuer is a bank.
}
