% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHistoricIssueData.R
\name{GetHistoricIssueData}
\alias{GetHistoricIssueData}
\title{Get historic issue data from DTCC}
\usage{
GetHistoricIssueData(start_date = "2010-01-01")
}
\arguments{
\item{start_date}{start-dato for uttrekk}
}
\description{
Get historic issue data from DTCC
}
\examples{
\dontrun{data = GetHistoricIssueData()}
}
