% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNewIssueData.R
\name{GetNewIssueData}
\alias{GetNewIssueData}
\title{Get new issue data from DTCC}
\usage{
GetNewIssueData(start_date = "2021-03-01")
}
\arguments{
\item{start_date}{start date of data extraction}
}
\description{
Get new issue data from DTCC
}
\examples{
\dontrun{new_data = GetNewIssueData()}

}
