% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagBanksCPCD.R
\name{TagBanksCPCD}
\alias{TagBanksCPCD}
\title{Tag banker i CP/CD-data fra DTCC}
\usage{
TagBanksCPCD(data)
}
\arguments{
\item{data}{datasett som skal tagges}
}
\description{
Tag banker i CP/CD-data fra DTCC
}
