% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCPCD.R
\name{AddMaturityBuckets}
\alias{AddMaturityBuckets}
\title{Adds maturity categories}
\usage{
AddMaturityBuckets(data)
}
\arguments{
\item{data}{(CP/CD) data set.}
}
\value{
Data set with column indicating whether issuer is a bank.
}
\description{
Adds a dummy indicating whether an CP/CD issuer is a bank.
}
