% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCPCD.R
\name{TagBanksCPCD}
\alias{TagBanksCPCD}
\title{Tags CP/CD issuers as bank/non-bank}
\usage{
TagBanksCPCD(data)
}
\arguments{
\item{data}{(CP/CD) data set.}
}
\value{
Data set with column indicating whether issuer is a bank.
}
\description{
Adds a dummy indicating whether an CP/CD issuer is a bank.
}
